make Rfc3986Escape a NOP in the common case
Rfc3986Escape now returns the input string immediately if nothing inside
it needs to be escaped.

This avoids doing the annoying allocations for the output buffer and its
bytes as well as the string copy at the end.

It also tracks the prefix of okay input bytes, adds them to the output
buffer, and avoids repeating the work of checking them again.